Refactor and remove form on index
