+++ views/index.jade
@@ -6,6 +6,3 @@ html( lang="en" )
 block content
   h1= title
   p Welcome to #{title}
-  form(action='/', method='post')
-     input(id='manual',name='manual',type='text',value='',placeholder='@')
-     input(type='submit',value='Submit')
